Struct jsonrpc_macros::Trailing

pub struct Trailing<T>(_);

A wrapper type without an implementation of Deserialize which allows a special implementation of Wrap for functions that take a trailing default parameter.

Methods

impl<T: DeserializeOwned> Trailing<T>

fn unwrap_or(self, other: T) -> T

Returns a underlying value if present or provided value.

fn unwrap_or_else<F: FnOnce() -> T>(self, f: F) -> T

Returns an underlying value or computes it if not present.

impl<T: Default + DeserializeOwned> Trailing<T>

fn unwrap_or_default(self) -> T

Returns an underlying value or the default value.